+++ src/Logs.hpp
@@ -5,4 +5,4 @@ void error();
 void stairs();
 void tree();
 
-void display(float value);
+void display(double value);
